Animal/Bird PreLanding: 96.86% -> 97.24% (narrow BOOL branch)

+++ src/Animal/Bird.cpp
@@ -154,7 +154,7 @@ BOOL TNerveAnimalBirdPreLanding::execute(TSpineBase<TLiveActor>* spine) const
 		                               scale * *angle, scale * *aware)
 		          != 0;
 	}
-	if (inSight) {
+	if ((u8)inSight) {
 		spine->pushAfterCurrent(&TNerveAnimalBirdGraphWander::theNerve());
 		return TRUE;
 	}
